+++ CliMeter/src/ch/uzh/ifi/rerg/se16_climeter/client/map/DataPoint.java
@@ -194,13 +194,4 @@ public class DataPoint {
 		}
 	}
 
-	/**
-	 * @pre -
-	 * @post -
-	 * @return the DataPoint as OverlayView
-	 */
-	public DataOverlay getDataOverlay() {
-		return this.dataPoint;
-	}
-
 }
