+++ CliMeter/test/ch/uzh/ifi/rerg/se16_climeter/client/filter/TimeLineTest.java
@@ -2,14 +2,8 @@ package ch.uzh.ifi.rerg.se16_climeter.client.filter;
 
 import static org.junit.Assert.*;
 
-import java.util.ArrayList;
-
 import org.junit.Test;
 import com.google.gwt.junit.client.GWTTestCase;
-import com.google.gwt.maps.client.LoadApi;
-import com.google.gwt.maps.client.LoadApi.LoadLibrary;
-
-import ch.uzh.ifi.rerg.se16_climeter.client.map.MapComposite;
 
 public class TimeLineTest extends GWTTestCase {
 
@@ -20,53 +14,37 @@ public class TimeLineTest extends GWTTestCase {
 
 	@Test
 	public void testTimeLine_null() {
-		// load all the libraries for use in the maps
-		ArrayList<LoadLibrary> loadLibraries = new ArrayList<LoadApi.LoadLibrary>();
-		loadLibraries.add(LoadLibrary.ADSENSE);
-		loadLibraries.add(LoadLibrary.DRAWING);
-		loadLibraries.add(LoadLibrary.GEOMETRY);
-		loadLibraries.add(LoadLibrary.PANORAMIO);
-		loadLibraries.add(LoadLibrary.PLACES);
-		loadLibraries.add(LoadLibrary.WEATHER);
-		loadLibraries.add(LoadLibrary.VISUALIZATION);
-
-		// thread with running map
-		Runnable onLoad = new Runnable() {
+		Filterable filterable = new Filterable() {
 			@Override
-			public void run() {
-				MapComposite mapComposite = new MapComposite();
-				TimeLine timeLine = new TimeLine(mapComposite, 1800, 2016);
-				assertNotNull(timeLine);
+			public void apply(Filter filter) {
+				// do something
+			}
+			@Override
+			public Filter getOldFilter() {
+				return new Filter();
 			}
 		};
 
-		LoadApi.go(onLoad, loadLibraries, true);
+		TimeLine timeLine = new TimeLine(filterable, 1800, 2016);
+		assertNotNull(timeLine);
 	}
 
 	@Test
 	public void testTimeLine_1() {
-		// load all the libraries for use in the maps
-		ArrayList<LoadLibrary> loadLibraries = new ArrayList<LoadApi.LoadLibrary>();
-		loadLibraries.add(LoadLibrary.ADSENSE);
-		loadLibraries.add(LoadLibrary.DRAWING);
-		loadLibraries.add(LoadLibrary.GEOMETRY);
-		loadLibraries.add(LoadLibrary.PANORAMIO);
-		loadLibraries.add(LoadLibrary.PLACES);
-		loadLibraries.add(LoadLibrary.WEATHER);
-		loadLibraries.add(LoadLibrary.VISUALIZATION);
-
-		// thread with running map
-		Runnable onLoad = new Runnable() {
+		Filterable filterable = new Filterable() {
 			@Override
-			public void run() {
-				MapComposite mapComposite = new MapComposite();
-				TimeLine timeLine = new TimeLine(mapComposite, 1800, 2016);
-				assertEquals(1800.0, timeLine.getMinValue());
-				assertEquals(2016.0, timeLine.getMaxValue());
+			public void apply(Filter filter) {
+				// do something
+			}
+			@Override
+			public Filter getOldFilter() {
+				return new Filter();
 			}
 		};
 
-		LoadApi.go(onLoad, loadLibraries, true);
+		TimeLine timeLine = new TimeLine(filterable, 1800, 2016);
+		assertEquals(1800.0, timeLine.getMinValue());
+		assertEquals(2016.0, timeLine.getMaxValue());
 	}
 
 }
